refractor style helper
